optimize

+++ DaGaunYun/CalcWageProject/code/project/build.sh
@@ -1,4 +1,10 @@
+if test -d build
+then
+	rm -rf build/*
+else
+	mkdir build
+fi
 
-rm -rf build/*
 cd build
 cmake ..
+make
